+++ debug.f08
@@ -405,7 +405,7 @@ contains
         !
         character(len=64) :: subroutine_name = 'debug_output'
 
-        real(4),      target,     intent(in) :: elem2R(:,:),  elemMR(:,:),  faceR(:,:)
+        real(8),      target,     intent(in) :: elem2R(:,:),  elemMR(:,:),  faceR(:,:)
         integer,   target,     intent(in) :: elem2I(:,:),  elemMI(:,:),  faceI(:,:)
         logical,   target,     intent(in) :: elem2YN(:,:), elemMYN(:,:), faceYN(:,:)
 
@@ -419,10 +419,10 @@ contains
 
         integer,       pointer ::  CI, UnitNumber
 
-        real(4),          pointer ::  thisdataR(:)
+        real(8),          pointer ::  thisdataR(:)
         integer,       pointer ::  thisdataI(:)
 
-        integer :: arrayContains= 0  ! =1 for real(4), 2 for integer, 3 for logical
+        integer :: arrayContains= 0  ! =1 for real(8), 2 for integer, 3 for logical
 
         integer, parameter :: dataR = 1
         integer, parameter :: dataI = 2
